// projects/ui-builder-component-lib/src/lib/model/popup-modal.ts
export interface PopupModal {
  icon?: any;
  title?: string;
  message?: string;
  cancelButton?: string;
  confirmButton?: string;
  checkboxText?: string;
  progBarCurrentStep?: number;
  progBarTotalSteps?: number;
  extendedMessage?: string;
  customClass?: string;
  backdropDismiss?: boolean;
}